add shared options and move to utils
