+++ src/utils/log.ts
@@ -1,4 +1,4 @@
-import type { Options } from './types'
+import type { Options } from '../types'
 
 type LogType = 
     | 'proxy'
